requirements.txt
fastapi==0.115.12
jinja2==3.1.6
python-dotenv==1.0.1
pydantic==2.7.3
requests==2.32.3
uvicorn==0.30.0